Fix issues and add improvements

+++ src/benchmarks/base.rs
@@ -265,16 +265,16 @@ where
 
 /// Format bytes to human-readable string.
 pub fn format_bytes(bytes: u64) -> String {
-    const KB: u64 = 1024;
-    const MB: u64 = KB * 1024;
-    const GB: u64 = MB * 1024;
+    const KB: f64 = 1024.0;
+    const MB: f64 = KB * 1024.0;
+    const GB: f64 = MB * 1024.0;
 
-    if bytes >= GB {
-        format!("{:.2} GB", bytes as f64 / GB as f64)
-    } else if bytes >= MB {
-        format!("{:.2} MB", bytes as f64 / MB as f64)
-    } else if bytes >= KB {
-        format!("{:.2} KB", bytes as f64 / KB as f64)
+    if bytes as f64 >= GB {
+        format!("{:.2} GB", bytes as f64 / GB)
+    } else if bytes as f64 >= MB {
+        format!("{:.2} MB", bytes as f64 / MB)
+    } else if bytes as f64 >= KB {
+        format!("{:.2} KB", bytes as f64 / KB)
     } else {
         format!("{} B", bytes)
     }
